6328449 acpica: create_d2a_subtree makes nodes for ACPI non-devices 6332549 acpica: update to Intel ACPI CA 20050930 source drop 6333936 vsnprintf: vsnprintf doesn't support 'h' or string-precision modifiers 6335392 psm: psm_common.c incorrectly uses ACPI resource end tags

+++ usr/src/uts/i86pc/io/acpica/namespace/nsparse.c
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Module Name: nsparse - namespace interface to AML parser
- *              $Revision: 9 $
+ *              $Revision: 1.10 $
  *
  *****************************************************************************/
 
@@ -141,7 +141,7 @@
 
 ACPI_STATUS
 AcpiNsOneCompleteParse (
-    UINT32                  PassNumber,
+    UINT8                   PassNumber,
     ACPI_TABLE_DESC         *TableDesc)
 {
     ACPI_PARSE_OBJECT       *ParseRoot;
